前端增加最愛等級

+++ ComicApi/Model/PageModel.cs
@@ -27,7 +27,7 @@ namespace ComicApi.Model
         public bool IsIgnore { get; set; }
         public bool IsFavorite { get; set; }
         // For Favorite Comic property
-        public int Level { get; set; }
+        public long Level { get; set; }
     }
 
     public class ComicModel
